requirements.txt
# Config
pyyaml

# Embeddings & vector store
ollama
chromadb

# HTTP (SPARQL queries to Fuseki)
requests

# API
fastapi
uvicorn[standard]
aiofiles
